refactor: restructured saveFleximarkCssToGlobalStorage command and related utilities
