+++ src/command/command/index.mts
@@ -1,4 +1,5 @@
 import collectAdmonitions from "./collectAdmonitions/index.mjs";
 import createNote from "./createNote/index.mjs";
+import resetGlobalFleximarkCss from "./resetGlobalFleximarkCss/index.mjs";
 
-export { collectAdmonitions, createNote };
+export { collectAdmonitions, createNote, resetGlobalFleximarkCss };
