+++ test/command/lib/css/index.test.mts
@@ -1,6 +1,6 @@
 import * as assert from "assert";
 
-import { getDefaultFleximarkCss } from "../../../src/command/css/index.mjs";
+import { getDefaultFleximarkCss } from "../../../../src/command/lib/css/index.mjs";
 
 export const suiteName = "fleximark.css Utility Tests";
 export const suite = () => {
